+++ apuntes-webapp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/ReviewService.java
@@ -6,7 +6,7 @@ import ar.edu.itba.paw.models.Review;
 
 public interface ReviewService {
 
-	public Review createReview(int fileid, int userid, double ranking, String review);
+	public Review createReview(int fileid, long userid, double ranking, String review);
 
 	public List<Review> findByFileId(int fileid);
